+++ config/metrics/averagequalityscore.php
@@ -1,5 +1,5 @@
 <?php
 return [
     "id" => "averagequalityscore",
-    "type" => "integer"
+    "type" => "decimal"
 ];
